Move duplicate parent resource key check to linting rule
