+++ common/changes/@cadl-lang/rest/parent-key-linting_2022-10-20-14-50.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@cadl-lang/rest",
+      "comment": "Move `duplicate-parent-key` check from the `parentResource` decorator to a linting rule",
+      "type": "patch"
+    }
+  ],
+  "packageName": "@cadl-lang/rest"
+}
