+++ packages/rest/lib/rest.cadl
@@ -1,5 +1,3 @@
-import "../dist/src/validate.js";
-
 import "./http.cadl";
 import "./resource.cadl";
