+++ packages/rest/src/index.ts
@@ -1,4 +1,5 @@
 export const namespace = "Cadl.Rest";
 export * as http from "./http/index.js";
+export * from "./linter.js";
 export * from "./resource.js";
 export * from "./rest.js";
